services/stt-live/requirements.txt
livekit-agents==0.12.8
livekit-plugins-silero==0.7.4
aiohttp==3.11.11
redis==5.2.1
